requirements.txt
fastapi == 0.111.0
uvicorn == 0.30.1
replit == 4.0.0
starlette == 0.37.2
uvicorn == 0.30.1
python-jose == 3.3.0
passlib == 1.7.4
sqlalchemy == 2.0.31
pydantic == 2.8.2
bcrypt == 4.0.1
FastAPI-SQLAlchemy  == 0.2.1
python-dotenv == 1.0.1
pytest == 8.2.2
psycopg2 == 2.9.9
alembic == 1.13.2
asyncpg == 0.29.0
minio == 7.2.7